Added an example to show how to use sud listeners.
